requirements.txt
discord.py
NHentai-API
pygelbooru
saucenao_api